Move non-window test to default_test
